pyproject.toml
[project]
name = "perplexica-mcp"
version = "0.3.0"
requires-python = ">=3.12"
dependencies = [
  "mcp[cli]",
  "httpx",
  "python-dotenv",
  "uvicorn",
]

[tool.uv]
